Moved annotation files into their ownpackages
