+++ tests/test_MeMoModel.py
@@ -10,10 +10,10 @@ import sys
 from pathlib import Path
 from src.MeMoMetabolite import MeMoMetabolite
 from src.MeMoModel import MeMoModel
-from src.annotateModelSEED import annotateModelSEED, annotateModelSEED_id
-from src.annotateChEBI import annotateChEBI
-from src.annotateBiGG import annotateBiGG, annotateBiGG_id
-from src.annotateAux import AnnotationResult
+from src.annotation.annotateModelSEED import annotateModelSEED, annotateModelSEED_id
+from src.annotation.annotateChEBI import annotateChEBI
+from src.annotation.annotateBiGG import annotateBiGG, annotateBiGG_id
+from src.annotation.annotateAux import AnnotationResult
 from src.removeDuplicateMetabolites import detectDuplicates, removeDuplicateMetabolites
 
 print(sys.version)
